% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEcosData.R
\name{ecosSearch}
\alias{ecosSearch}
\title{Search for ECOS statistical codes}
\usage{
ecosSearch(x)
}
\arguments{
\item{x}{A character vector containing search keywords}
}
\value{
A tibble containing the matching statistical codes
}
\description{
This function searches for ECOS statistical codes in a preloaded dataset by matching keywords provided by the user.
}
