% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBisData.R
\name{getBisData}
\alias{getBisData}
\title{Download and extract a BIS database file}
\usage{
getBisData(url, ...)
}
\arguments{
\item{url}{the URL of the BIS database ZIP file to be downloaded (obtain from \code{\link[=getBisDB]{getBisDB()}})}

\item{...}{additional arguments passed to \code{\link[utils:download.file]{utils::download.file()}} (e.g. \code{quiet = TRUE})}
}
\value{
a data frame with the contents of the downloaded CSV file
}
\description{
This function downloads and extracts a BIS database file, then reads it into R as a CSV file.
}
\examples{
\dontrun{
db <- getBisDB()
# Download the first available dataset
df <- getBisData(db$url[1])
head(df)
}
}
