% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMFdata-package.R
\docType{package}
\name{getMFdata-package}
\alias{getMFdata}
\alias{getMFdata-package}
\title{getMFdata: Retrieving data from BOK, IMF, OECD, BIS, etc. through official APIs}
\description{
This package retrieves official statistics from the Bank of Korea (BOK) and Financial Supervisory Service (FSS) in South Korea and international organisation including OCED, BIS and IMF.
}
\author{
\strong{Maintainer}: Junho Lee \email{morner@fss.or.kr}

}
\keyword{internal}
