% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usdkrw}
\alias{usdkrw}
\title{USDKRW exchange rate data}
\format{
\subsection{\code{usdkrw}}{

A data frame with 410 rows and 2 columns:
\describe{
\item{yearmon}{year.month as a numeric}
\item{usdkrw3}{exchange rate}
}
}
}
\source{
\url{https://ecos.bok.or.kr/}
}
\usage{
usdkrw
}
\description{
The average monthly exchange rate US DOLLAR (USD) to SOUTH KOREAN WON (KRW) since Jan. 1990.
}
\keyword{datasets}
