% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{EcosTerm}
\alias{EcosTerm}
\title{EcosTerm function}
\usage{
EcosTerm(time, type = c("A", "Q", "M", "D"))
}
\arguments{
\item{time}{A time object.}

\item{type}{A character string that specifies the desired output format. It must be one of the following: "A" for annual, "Q" for quarterly, "M" for monthly, or "D" for daily. Default is "A".}
}
\value{
A character string containing the time value in the specified format.
}
\description{
This function returns a character string containing a time value in a specified format.
}
