% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBisData.R
\name{getBisData}
\alias{getBisData}
\title{Download and extract a BIS database file}
\usage{
getBisData(url, ...)
}
\arguments{
\item{url}{the URL of the BIS database file to be downloaded}

\item{...}{other arguments delivered to download.file function}
}
\value{
a data frame with the contents of the downloaded file
}
\description{
This function downloads and extracts a BIS database file, then reads it into R as a csv file.
}
