% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBisData.R
\name{getBisDB}
\alias{getBisDB}
\title{Function to extract data from Bank for International Settlements database}
\usage{
getBisDB()
}
\value{
a tibble with the name and URL of each available dataset in the BIS database
}
\description{
This function extracts data from the Bank for International Settlements database and returns it as a tibble.
}
