% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECOSdata.R
\name{getEcosCode}
\alias{getEcosCode}
\title{Retrieve statistical codes from ECOS database}
\usage{
getEcosCode(ECOS_key = ECOS_key, STAT_CODE, lang = "kr")
}
\arguments{
\item{ECOS_key}{An ECOS API key}

\item{STAT_CODE}{A statistical code}

\item{lang}{The language of the returned data (default: "kr")}
}
\value{
A tibble containing the retrieved statistical codes
}
\description{
This function retrieves statistical codes from the ECOS database by sending a GET request to the ECOS API.
}
