% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECOSdata.R
\name{getEcosData}
\alias{getEcosData}
\title{Function to retrieve data from Bank of Korea's ECOS system}
\usage{
getEcosData(
  ECOS_key,
  stat_code,
  period,
  start_time,
  end_time,
  item_code1,
  item_code2,
  item_code3,
  lang = "kr"
)
}
\arguments{
\item{ECOS_key}{character string containing user's unique API key}

\item{stat_code}{character string containing specific statistical code}

\item{period}{character string containing period of data (ex. 'DD', 'MM')}

\item{start_time}{character string containing start date of data (YYYYMMDD)}

\item{end_time}{character string containing end date of data (YYYYMMDD)}

\item{item_code1}{character string containing first item code}

\item{item_code2}{character string containing second item code}

\item{item_code3}{character string containing third item code}

\item{lang}{character string containing language of data ('kr' for Korean, 'en' for English)}
}
\value{
data frame containing TIME and DATA_VALUE columns of requested data
}
\description{
This function utilizes Bank of Korea's ECOS API to retrieve statistical data
based on user input.
}
