% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getECOSdata.R
\name{getEcosList}
\alias{getEcosList}
\title{retrieving stats list from ECOS}
\usage{
getEcosList(ECOS_key = ECOS_key, lang = "kr")
}
\arguments{
\item{ECOS_key}{ECOS API key}

\item{lang}{language code (default is 'kr')}
}
\value{
A tibble containing the statistics table list
}
\description{
retrieving stats list from ECOS
}
