% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFisisData.R
\name{getFsisInfos}
\alias{getFsisInfos}
\title{Function to retrieve FSIS information}
\usage{
getFsisInfos(
  api_key,
  info_name = c("companySearch", "statisticsListSearch", "accountListSearch"),
  item_code = "A"
)
}
\arguments{
\item{api_key}{FSIS API key}

\item{info_name}{Character string. Name of information to retrieve. Choices are 'companySearch', 'statisticsListSearch', or 'accountListSearch'.}

\item{item_code}{Character string. Code for the search parameter. Default is 'A'.}
}
\value{
A list with the retrieved information.
}
\description{
Get information from FSIS API.
}
