% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImfData.R
\name{getImfData}
\alias{getImfData}
\title{Function to retrieve financial data from the International Monetary Fund (IMF)}
\usage{
getImfData(indicator_code, countries, start_date, end_date, frequency)
}
\arguments{
\item{indicator_code}{Indicator code for the financial data to be retrieved}

\item{countries}{List of ISO country codes for the countries to be included in the data}

\item{start_date}{Starting date for the data (in the format "YYYY-MM-DD")}

\item{end_date}{Ending date for the data (in the format "YYYY-MM-DD")}

\item{frequency}{Time frequency for the data (e.g. "A" for annual, "Q" for quarterly)}
}
\value{
a list with the requested financial data
}
\description{
This function retrieves financial data from the International Monetary Fund (IMF) using the IMF Data Services API.
}
