% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOecdData.R
\name{getOecdDB}
\alias{getOecdDB}
\title{Get OECD Database}
\usage{
getOecdDB(url)
}
\arguments{
\item{url}{The URL of the OECD database API}
}
\value{
a list containing the structure and dataSets of the OECD database
}
\description{
This function retrieves data from an OECD database, formatted as a JSON object,
given a specific starting time.
}
