% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImfData.R
\name{processImfData}
\alias{processImfData}
\title{Process IMF data retrieved from the IMF API}
\usage{
processImfData(IMFdata)
}
\arguments{
\item{IMFdata}{A list of financial data retrieved from the IMF API using the getImfdata function}
}
\value{
a data frame with the processed IMF data
}
\description{
This function processes the financial data retrieved from the IMF API using the getImfdata function.
}
